Moved LevelRewards to the Service directory
